Split helpers.py and rename. Closes #177.
